+++ R/workability.R
@@ -1,6 +1,10 @@
 #' Calculate indicator for workability 
 #'
-#' This function calculates the workability of soils, given as potential length of the growing season. Based on Huinink (2018)
+#' This function calculates the workability of soils, given as a value of relative season length between 0 and 1. 
+#' A relative season length of 1 indicates that the water table is sufficiently low for the soil to be workable for the entire growing season required by the crop.
+#' The required ground water table for workability is determined by soil type and soil properties. Hydrological variables determine the groundwater table for each day of the year.
+#' The option calcyieldloss allows for calculation of yield loss based on the relative season length, differentiating in yield loss between six groups of crops
+#' Based on Huinink (2018)
 #' 
 #' @param A_CLAY_MI (numeric) The clay content of the soil (\%)
 #' @param A_SILT_MI (numeric) The silt content of the soil (\%)
@@ -8,7 +12,7 @@
 #' @param B_SOILTYPE_AGR (character) The agricultural type of soil
 #' @param B_GWL_GLG (numeric) The lowest groundwater level averaged over the most dry periods in 8 years in cm below ground level
 #' @param B_GWL_GHG (numeric) The highest groundwater level averaged over the most wet periods in 8 years in cm below ground level
-#' @param B_Z_TWO  (numeric) The distance between ground level and groundwater level at which the groundwater can supply the soil surface with 2mm water per day (in cm)
+#' @param B_GWL_ZCRIT  (numeric) The distance between ground level and groundwater level at which the groundwater can supply the soil surface with 2mm water per day (in cm)
 #' @param calcyieldloss (boolean) whether the function includes yield loss, options: TRUE or FALSE (default).
 #' 
 #' @import data.table
@@ -17,7 +21,7 @@
 #'  
 #' @export
 calc_workability <- function(A_CLAY_MI, A_SILT_MI, B_LU_BRP, B_SOILTYPE_AGR, 
-                             B_GWL_GLG, B_GWL_GHG, B_Z_TWO,
+                             B_GWL_GLG, B_GWL_GHG, B_GWL_ZCRIT,
                              calcyieldloss = FALSE) {
   
   # define variables used within the function
@@ -42,7 +46,7 @@ calc_workability <- function(A_CLAY_MI, A_SILT_MI, B_LU_BRP, B_SOILTYPE_AGR,
 
   # Check inputs
   arg.length <- max(length(A_CLAY_MI), length(A_SILT_MI), length(B_LU_BRP), length(B_SOILTYPE_AGR), 
-                    length(B_GWL_GLG), length(B_GWL_GHG), length(B_Z_TWO))
+                    length(B_GWL_GLG), length(B_GWL_GHG), length(B_GWL_ZCRIT))
   checkmate::assert_numeric(A_CLAY_MI, lower = 0, upper = 100, any.missing = FALSE, len = arg.length)
   checkmate::assert_numeric(A_SILT_MI, lower = 0, upper = 100, any.missing = FALSE, len = arg.length)
   checkmate::assert_numeric(B_LU_BRP, any.missing = FALSE, min.len = 1, len = arg.length)
@@ -52,7 +56,7 @@ calc_workability <- function(A_CLAY_MI, A_SILT_MI, B_LU_BRP, B_SOILTYPE_AGR,
   checkmate::assert_numeric(B_GWL_GLG, lower = 0, any.missing = FALSE, len = arg.length)
   checkmate::assert_numeric(B_GWL_GHG, lower = 0, any.missing = FALSE, len = arg.length)
   checkmate::assert_true(all(B_GWL_GHG < B_GWL_GLG))
-  checkmate::assert_numeric(B_Z_TWO, lower = 0, upper = 400, any.missing = FALSE, len = arg.length)
+  checkmate::assert_numeric(B_GWL_ZCRIT, lower = 0, upper = 400, any.missing = FALSE, len = arg.length)
 
   # Collect in data table
   dt <- data.table(id = 1:arg.length,
@@ -62,13 +66,13 @@ calc_workability <- function(A_CLAY_MI, A_SILT_MI, B_LU_BRP, B_SOILTYPE_AGR,
                    B_SOILTYPE_AGR = B_SOILTYPE_AGR,
                    B_GWL_GLG = B_GWL_GLG,
                    B_GWL_GHG = B_GWL_GHG,
-                   B_Z_TWO = B_Z_TWO)
+                   B_GWL_ZCRIT = B_GWL_ZCRIT)
   
   # merge with OBIC crop and soil table
   dt <- merge(dt, crops.obic[, list(crop_code, crop_waterstress, crop_season)], 
               by.x = "B_LU_BRP", by.y = "crop_code")
   dt <- merge(dt, soils.obic[, list(soiltype, soiltype.m)], by.x = "B_SOILTYPE_AGR", by.y = "soiltype")
-  dt <- merge(dt, season.obic, by.x = 'crop_season', by.y = 'landuse')
+  dt <- merge(dt, season.obic, by.x = c('crop_season','soiltype.m'), by.y = c('landuse', 'soiltype.m'))
   
   ## determine workability key numbers
 
@@ -101,7 +105,7 @@ calc_workability <- function(A_CLAY_MI, A_SILT_MI, B_LU_BRP, B_SOILTYPE_AGR,
   dt[, rdh_fall := gws_sub_workingdepth]
   
   # test 2: At what groundwater level is  capillary rise lower than evaporation (<2mm/d), required depth capilairy abbreviated as rdc
-  dt[, rdc := B_Z_TWO]
+  dt[, rdc := B_GWL_ZCRIT]
 
   # Choose lowest required depth as required depth
   dt[,rd_spring := fifelse(rdh_spring <= rdc, rdh_spring,rdc)]
@@ -173,6 +177,9 @@ calc_workability <- function(A_CLAY_MI, A_SILT_MI, B_LU_BRP, B_SOILTYPE_AGR,
     value <- dt[,yield]
     
   } else {
+    
+    # add temporaly check
+    dt[rsl<0, rsl := 0]
     
     # return relative length season
     value <- dt[,rsl]