+++ NEWS.md
@@ -1,5 +1,8 @@
 # Changelog OBIC
 
+## Version 1.0.3.9000 DEVELOPMENT
+### Removed
+* redundant parameters A_CACO3 and A_COM_FR from required inputs of wrapper functions 
 ## UNRELEASED
 ### Added
 * Adds GitHub Action to run R-CMD-Check for changes on `master` and `development`
@@ -14,15 +17,68 @@
 * ppr_bouwplan_tables.r in dev/scripts
 * merge_nema_tables.r in dev/scripts
 
-## Version 1.1.0 2021-05-03
+## Version 1.0.2.9000 DEVELOPMENT
 ### Added
-* argument B_GWL_GHG, B_GWL_GLG and B_Z_TWO to obic_field and obic_field_dt
+* script ppr_bouwplan_tables.r in dev/scripts
+* script merge_nema_tables.r in dev/scripts
+* script ppr_crops_crumbleability.r in dev/scripts
+* arguments B_GWL_GHG, B_GWL_GLG and B_Z_TWO to obic_field and obic_field_dt
 * function calc_workability is active now
+* function calc_makkink to add makking factors to crop table
+* function calc_psp to estimate precipitation surplus
+* functions ind_nematodes and ind_nematodes_list to estimate the index for nematode risks.
+* function ind_gw_recharge to estimate the groundwater recharge index
+* functions calc_n_efficiency and ind_n_efficiency to estimate N efficiency
+* functions calc_pesticide_leaching and ind_pesticide_leaching to evaluate retention of pesticides in topsoil
+* check to ppr_weight_obic to check if all indicators in weight.obic occur in column_descriptions_obic
+* vignette obic_workability, describing how workability is calculated and affected by its input variables
+* vignette obic_introduction as a short tutorial describing the OBIC principles
+* vignette obic_score_aggregation to illustrate the aggregation principles used
+* table weather.obic added with montly mean precipitation, temperature and evaporation
+* table crops.makkink added with mean Makkink factors per crop type
+* table nema.crop.rot.obic added with crop sensitivity for nematodes
+* table nema.obic with treshold values per species
 
 ### Changed
-* updated documentation, weight.obic for workability
+* update documentation for workability
+* update weight.obic for workability
+* update column_descriptions_obic.Rdata
+* update documentation for nematodes
 * update calc_workability
-* update column_descriptons_obic.Rdata
+* renamed B_Z_TWO to B_GWL_ZCRIT
+* improved function description of ind_workability
+* increased required days post GLG for maize on sand or loess to match October 1
+* increased required days post GLG for maize on other soils to match October 20
+* season.obic can now be merged on both landuse and soiltype.m
+* update binnenveld with BodemSchat data (via OBIC-helpers)
+* obic_field can be used for multiple fields with field-ID
+
+### Fixed
+* error in grass_age: age was incorrectly estimated for multiple fields
+* removed negative values from season.obic
+* ensured total_days >= req_days_pre_glg + req_days_post_glg in season.obic 
+* root depth was incorrectly assigned for maize and nature 
+
+## Version 1.0.2 2021-05-20
+### Added
+* ind_man_ess.R and tests are added for update aggregated management scores, issue #131
+
+### Changed
+* obic_field() makes use of ind_man_ess
+
+### Fixed
+* check B_GWL_CLASS in calc_waterstressindex(), issue #129
+
+## Version 1.0.1 2021-05-05
+### Added
+* management.obic table added to link management measures to ecosystem services
+* calc_man_ess.R and tests are added to estimate grouped impact of management on ecosystem servcices, issue #125
+
+### Changed
+* I_M_* and I_*_BCS indicators for management measures and BCS are removed, issue #126
+* values -999 are replaced by NA 
+* column_descriptons_obic.Rdata is updated
+* weight.obic is updated
 
 ## Version 1.0.0 2021-04-28
 ### Added